+++ packages/patrol_cli/lib/src/android/android_test_backend.dart
@@ -45,7 +45,7 @@ class AndroidTestBackend {
   late final String? javaPath;
 
   Future<void> build(AndroidAppOptions options) async {
-    await buildApkConfigOnly(options.flutter.command);
+    await buildApkConfigOnly(options.flutter);
     await loadJavaPathFromFlutterDoctor(options.flutter.command);
     await detectOrchestratorVersion(options);
 
@@ -65,7 +65,10 @@ class AndroidTestBackend {
               ),
               runInShell: true,
               workingDirectory: _rootDirectory.childDirectory('android').path,
-              environment: javaPath != null ? {'JAVA_HOME': javaPath!} : {},
+              environment: switch (javaPath) {
+                final String javaPath => {'JAVA_HOME': javaPath},
+                _ => {},
+              },
             )
             ..disposedBy(scope);
       process.listenStdOut((l) => _logger.detail('\t: $l')).disposedBy(scope);
@@ -90,7 +93,10 @@ class AndroidTestBackend {
               ),
               runInShell: true,
               workingDirectory: _rootDirectory.childDirectory('android').path,
-              environment: javaPath != null ? {'JAVA_HOME': javaPath!} : {},
+              environment: switch (javaPath) {
+                final String javaPath => {'JAVA_HOME': javaPath},
+                _ => {},
+              },
             )
             ..disposedBy(scope);
       process.listenStdOut((l) => _logger.detail('\t: $l')).disposedBy(scope);
@@ -162,13 +168,21 @@ class AndroidTestBackend {
   /// Execute `flutter build apk --config-only` to generate the gradlew file.
   ///
   /// This fix issue: https://github.com/leancodepl/patrol/issues/1668
-  Future<void> buildApkConfigOnly(FlutterCommand flutterCommand) async {
+  Future<void> buildApkConfigOnly(FlutterAppOptions options) async {
     final process = await _processManager.start([
-      flutterCommand.executable,
-      ...flutterCommand.arguments,
+      options.command.executable,
+      ...options.command.arguments,
       'build',
       'apk',
       '--config-only',
+      if (options.buildName case final buildName?) ...[
+        '--build-name',
+        buildName,
+      ],
+      if (options.buildNumber case final buildNumber?) ...[
+        '--build-number',
+        buildNumber,
+      ],
       '-t',
       'integration_test/test_bundle.dart',
     ], runInShell: true);
@@ -270,7 +284,7 @@ class AndroidTestBackend {
               runInShell: true,
               environment: {
                 'ANDROID_SERIAL': device.id,
-                ...javaPath != null ? {'JAVA_HOME': javaPath!} : {},
+                if (javaPath case final javaPath?) ...{'JAVA_HOME': javaPath},
               },
               workingDirectory: _rootDirectory.childDirectory('android').path,
             )